Fixed some with CSS
