+++ code/src/App.css
@@ -10,8 +10,8 @@
   .app {
     width: 50vh;
     height: 60vh;
-    max-width: 450px;
+    max-width: 400px;
     min-height: 600px;
-    margin-top: 40px;
+    margin-top: 60px;
   }
 }
