+++ code/src/components/Header.js
@@ -1,11 +1,11 @@
-import React from 'react'
-import './Header.css'
+import React from 'react';
+import './Header.css';
 
 export const Header = () => {
-    return (
-        <header className="myheader">
-            <h1>To-do-list</h1>
-            <p>This is a to-do list. Enter what you need to do and check it off once it's finished!</p>
-        </header>
-    )
-}
+  return (
+    <header className="myheader">
+      <h1>To-do-list</h1>
+      <p>This is a to-do list. Enter what you need to do and check it off once it's finished!</p>
+    </header>
+  );
+};
