+++ code/src/components/TodoList.css
@@ -41,12 +41,15 @@
   padding: 15px 15px;
   margin-bottom: 10px;
   box-shadow: rgba(60, 64, 67, 0.3) 0px 1px 2px 0px, rgba(60, 64, 67, 0.15) 0px 2px 6px 2px;
-  position: relative;
+  display:flex;
+  width:100%;
+  align-items: center;
 }
 
-.item:hover {
-  background-color: #eeeeec;
-  transition: 0.2s;
+.todo-item-text {
+  display:flex;
+  flex-direction: column;
+  flex-grow: 2;
 }
 
 .item span {
@@ -58,38 +61,34 @@
 }
 
 .delete {
-  padding:5px 25px;
-  width:100px;
+  padding:10px 15px;
+  width:80px;
   border:none;
   background:#ac4646;
   color:white;
   border-radius: 5px;
   font-weight: 700;
-  position:absolute;
-  right:10px;
-  top:10px;
+  align-self: flex-end;
+}
+
+.delete:hover {
+  background-color:#b47979;
+  transition: .2s;
 }
 
 .update {
-  padding:5px 25px;
-  width:100px;
   border:none;
-  background:#6fac46;
-  color:white;
-  border-radius: 5px;
-  font-weight: 700;
-  position:absolute;
-  right:10px;
-  bottom:10px;
+  margin-right:20px;
 }
 
 .date {
   font-size: 12px;
-  margin-top: 10px;
+  margin:0;
 }
 
 .todo-header {
   font-weight: 600;
+  margin:0;
 }
 
 @media only screen and (min-width: 599px) {
@@ -105,9 +104,5 @@
 
   .list-of-items-container::-webkit-scrollbar {
     display: none;
-  }
-
-  .date {
-    margin-top: 20px;
   }
 }
